# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LifeOS Backend API
  version: 2.0.0
  description: |
    LifeOS orchestrates Gmail, AI extraction, conflict detection, and Google Calendar publishing.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /api/demo/conflicts/{id}/resolve:
    post:
      tags: [Demo]
      summary: Resolve a demo conflict
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MailResolutionRequest"
      responses:
        "200":
          description: Conflict resolution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseResolve"
...
  /api/mail/conflicts/{id}/resolve:
    post:
      tags: [Mail]
      summary: Resolve a conflict
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MailResolutionRequest"
      responses:
        "200":
          description: Resolution response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MailConflictResolveResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiMeta:
      type: object
      properties:
        requestId:
          type: string
        generatedAt:
          type: string
          format: date-time
...
    ApiResponseResolve:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/MailConflictResolveResponse"
        meta:
          $ref: "#/components/schemas/ApiMeta"
...
    MailResolutionRequest:
      type: object
      properties:
        optionId:
          type: string
        delegateToEmail:
          type: string
        notifyRecipientEmail:
          type: string
      required: [optionId]
    MailConflictResolveResponse:
      type: object
      properties:
        conflictId:
          type: string
        status:
          type: string
        selectedOptionId:
          type: string
        updatedEvents:
          type: array
          items:
            $ref: "#/components/schemas/EventResponse"
        notifications:
          type: array
          items:
            type: object
            additionalProperties: true
        resolvedAt:
          type: string
          format: date-time
        calendarUpdated:
          type: boolean
...
    EventResponse:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        status:
          type: string
        confidence:
          type: number
...